+++ infra/variables.tf
@@ -53,3 +53,21 @@ variable "alert_from_email" {
   type        = string
   default     = "gclark0812@gmail.com"
 }
+
+variable "banshee_web_password" {
+  description = "Password for the Banshee web interface"
+  type        = string
+  sensitive   = true
+}
+
+variable "web_password" {
+  description = "Password for the Banshee web interface"
+  type        = string
+  sensitive   = true
+}
+
+variable "alert_recipients" {
+  description = "List of email addresses to receive alerts"
+  type        = list(string)
+  default     = ["gclark0812@gmail.com", "psmith1111@icloud.com"]
+}